refactor: narrow the KMS wrapper to only the tracked CompletionException leak
The prior getAead-level URI validation (doesSupport + ARN-segment check)
was the wrong abstraction: it duplicated a piece of Tink's own private
URI parsing to work around a bug that belongs to tink-java-awskms#9, not
to a generic KmsClient wrapper. Removes it entirely -- getAead now calls
delegate.getAead directly and does not catch anything from it. Malformed
key URIs are the caller's problem to validate where they're loaded, not
this wrapper's.

Renamed ExceptionTranslatingKmsClient to
CompletionExceptionTranslatingKmsClient so its one remaining
responsibility -- unwrapping a CompletionException from encrypt/decrypt
-- is explicit in the name. It is now generic over any KmsClient again,
with nothing AWS- or Tink-specific in its logic.

Verified the new getAead-propagates-unchanged test actually guards this:
temporarily reintroduced a broad catch and confirmed exactly that test
fails.
